requirements.txt
praw==7.7.0
spacy==3.7.2
transformers==4.44.2
torch==2.4.1
streamlit==1.38.0
python-dotenv==1.0.1
google-generativeai==0.8.5
